LPS-X Make CustomJspBagRegistryUtil unit test
