--- portal-impl/test/integration/com/liferay/portal/deploy/hot/bundle/customjspbagregistryutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for CustomJspBagRegistryUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
